+++ cpu/atmega128rfa1/include/periph_cpu.h
@@ -26,6 +26,16 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
+/**
+ * @name    Power management configuration
+ * @{
+ */
+#define PM_NUM_MODES            (5)
+#define AVR8_PM_SLEEP_MODE_0    SLEEP_MODE_PWR_DOWN /**< Power Down */
+#define AVR8_PM_SLEEP_MODE_1    SLEEP_MODE_PWR_SAVE /**< Power Save */
+#define AVR8_PM_SLEEP_MODE_2    SLEEP_MODE_STANDBY  /**< Standby */
+#define AVR8_PM_SLEEP_MODE_3    SLEEP_MODE_ADC      /**< Sleep ADC low noise */
+/** @} */
 
 /**
  * @name   Available ports on the ATmega128rfa1 MCU
