+++ cpu/atxmega/Kconfig
@@ -19,7 +19,6 @@ config CPU_COMMON_ATXMEGA
     select HAS_PERIPH_GPIO
     select HAS_PERIPH_GPIO_IRQ
     select HAS_PERIPH_NVM
-    select HAS_PERIPH_PM
     select HAS_PERIPH_TIMER
     select HAS_PERIPH_TIMER_PERIODIC
