+++ cpu/atxmega/Makefile.features
@@ -11,7 +11,6 @@ FEATURES_PROVIDED += cpu_core_atxmega
 FEATURES_PROVIDED += periph_cpuid
 FEATURES_PROVIDED += periph_gpio periph_gpio_irq
 FEATURES_PROVIDED += periph_nvm
-FEATURES_PROVIDED += periph_pm
 FEATURES_PROVIDED += periph_timer periph_timer_periodic
 
 # Add atxmega configurations. This configuration enables modules that are only available when
