--- cpu/atxmega/atxmega.config
@@ -1,2 +1,0 @@
-# All ATxmega based CPUs provide PM
-CONFIG_MODULE_PM_LAYERED=y
