+++ cpu/atxmega/periph/pm.c
@@ -19,8 +19,6 @@
  * @}
  */
 
-#include <avr/sleep.h>
-
 #include "cpu_pm.h"
 #include "irq.h"
 #include "periph/pm.h"
@@ -71,44 +69,6 @@ void pm_reboot(void)
     /* Disable CCP for Protected IO registerand set new value*/
     _PROTECTED_WRITE(RST_CTRL, RST_SWRST_bm);
     while (1) {}
-}
-
-/*
- * DEBUG may affect this routine.
- *
- * --- Do NOT add DEBUG macro here ---
- *
- */
-void pm_set(unsigned mode)
-{
-    unsigned irq_state = irq_disable();
-
-    if (avr8_is_uart_tx_pending() && mode < 4) {
-        irq_restore(irq_state);
-        return;
-    }
-
-    switch (mode) {
-    case 0:
-        set_sleep_mode(SLEEP_SMODE_PDOWN_gc);
-        break;
-    case 1:
-        set_sleep_mode(SLEEP_SMODE_PSAVE_gc);
-        break;
-    case 2:
-        set_sleep_mode(SLEEP_SMODE_STDBY_gc);
-        break;
-    case 3:
-        set_sleep_mode(SLEEP_SMODE_ESTDBY_gc);
-        break;
-    default:
-        set_sleep_mode(SLEEP_SMODE_IDLE_gc);
-    }
-    sleep_enable();
-    sei();
-    sleep_cpu();
-    sleep_disable();
-    irq_restore(irq_state);
 }
 
 void pm_periph_enable(pwr_reduction_t pwr)
