+++ cpu/avr8_common/Kconfig
@@ -10,6 +10,7 @@ config CPU_ARCH_AVR8
     bool
     select HAS_ARCH_8BIT
     select HAS_ARCH_AVR8
+    select HAS_PERIPH_PM
 
     select MODULE_MALLOC_THREAD_SAFE if TEST_KCONFIG
     select MODULE_TINY_STRERROR_AS_STRERROR if TEST_KCONFIG
