+++ cpu/avr8_common/Makefile
@@ -2,6 +2,6 @@
 MODULE = avr8_common
 
 # add a list of subdirectories, that should also be build
-DIRS = avr_libc_extra
+DIRS = periph avr_libc_extra
 
 include $(RIOTBASE)/Makefile.base
