+++ cpu/avr8_common/Makefile.features
@@ -1,3 +1,4 @@
 FEATURES_PROVIDED += arch_8bit
 FEATURES_PROVIDED += arch_avr8
 FEATURES_PROVIDED += cpp
+FEATURES_PROVIDED += periph_pm
