+++ cpu/avr8_common/avr8_common.config
@@ -1,0 +1,2 @@
+# All AVR-8 based CPUs provide PM
+CONFIG_MODULE_PM_LAYERED=y
