+++ cpu/avr8_common/periph/Makefile
@@ -1,0 +1,3 @@
+MODULE = avr8_common_periph
+
+include $(RIOTMAKE)/periph.mk
